% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_haiexplorer}
\alias{launch_haiexplorer}
\title{Launch the HAI Explorer Shiny App}
\usage{
launch_haiexplorer()
}
\value{
Starts a Shiny app in your default web browser.
}
\description{
This function launches the interactive Shiny application included in the
BHAIpkgHafsa package. The app allows users to explore healthcare-associated
infection (HAI) burden data by infection type with interactive plots and tables.
}
\examples{
if (interactive()) {
  launch_haiexplorer()
}
}
